Bugfix: Saved pointer wasn't restored.
This made the two pointers running in sync, degrading the capability of
detecting addressing flaws (likely impossible to actually test as "clean"
address line shorts or breaks would still be detected).

It also causes the RAM test to test cells beyond the end of RAM, but
normally there the beginning of the RAM should mirror, hiding this.