Bugfix: Made fixes towards passing basic tests.
An emulator bug was discovered in the handling of the Half-Carry flag in
instructions taking a carry (it doesn't behave according to a ripple
subtracter while the real hardware does as expectable, documentations are
difficult to follow in this regard as they don't specify the behavior as
a ripple subtracter).

The emulator has to be fixed first before continuing development.