Fine-tuned the multiplication test, and updated documentation.
The goal of the fine-tuning was to set up the test so its first pass can
detect the most common probable failure modes of the chip.

The documentations were updated to reflect the current stage of
development: The library is almost completed, testing almost every
instruction in an adequate manner.